Menambahkan step wizard pada dialog mahasiswa baru
