+++ package.json
@@ -21,6 +21,7 @@
     "d3-time": "^3.0.0",
     "d3-time-format": "^4.0.0",
     "formik": "^2.2.6",
+    "lodash": "^4.17.21",
     "moment": "^2.29.1",
     "query-string": "^7.0.0",
     "react": "^17.0.2",
